Make ReSt support more robust.
Search for a rst2html or rst2html.py to find the likely desired script, and die if it cannot be found. Then open more safely, with nicer Windows support, and the "generator" meta element removed (why is `--no-generator` ignored??). And also return `undef` if there appears to be no actual content.