Fixed code-block directive with Docutils >= 0.9
Docutils 0.9 tries to apply syntax highlight (previously this was only
available into Sphinx) failing if Pygments is not available. We
unconditionally disable syntax highlight to make the test more
predictable (highlight would require the proper stylesheet to be used
anyway).

The test now passes for Docutils 0.9 and 0.10. It fails with Docutils
0.8 but the output is correct anyway. 0.7 is broken instead: will fix in
another patch.

Fixes issue #6.