Add `raw` option where possible [#2]
This adds a `raw` option that can be passed to `parse()` that causes the
parser to return the raw output of the underlying layer, without
wrapping in the minimal HTML skeleton.

This is only possible for a subset of the existing parsers; this patch
makes no attempt to remove the HTML wrappers in the cases where that
wrapper is being added by the underlying parser.

Note also that this patch does not add any tests for this new behavior,
because there don't appear to be any cases where input/output
transformations are actually being currently tested.