Add GitHub workflows and fix Windows issues
Remove the Travis CI config and add GitHub workflows to test and
release. Test on Perl 5.8 - 5.38 on Linux, macOS, and Windows. Modify
the formats test with an environment variable to require tests to run
for all formats or for specific formats.

Fix Asciidoc and Rest syntax errors on Windows, and update Asciidoc to
find `.bat` and `.exe` files on Windows. The output of `asciidoc` on
Windows is not UTF-8 clean somehow, so add a filter to match the invalid
output of ü on Windows. Also move the replacement for the change in the
Rest xml declaration into a filter.

Add actions to Build.PL to output changes for GitHub releases and to
print the tarball name, as needed for the release workflow.

Moved the Pod testing dependencies from the Build.PL `recommends` list
to the `develop` list, via the `meta_merge` parameter.