Prefer Asciidoctor to asciidoc
Asciidoctor is the reference implementation now, for the latest Asciidoc
spec, which is now referenced from the docs. We still fall back on
`asciidoc` when we can't find `asciidoctor`.

Update the tests to detect which client has been found and compare the
output as appropriate. This ensures that we can test both properly. Then
install `asciidoctor` in the GitHub workflow after tests have run for
`asciidoc`, so we can be sure to properly test both. Use the
`TEXT_MARKUP_TEST_ASCIIDOC` environment variable to specify which to
test for, so we are sure to test both.

Move the Asciidoc CLI detection into a subroutine so that the test can
use it to detect which CLI was found and compar it to the proper output
file.

I wanted to also fall back on `asciidoc3` after `asciidoctor` and before
`asciidoc`, but the `pip install`ed version on my Mac could not run
because it couldn't find its configuration file. So leave it out for
now.

While at it, remove the custom Rest and Asciidoc path searching on
Windows; it was quite dated --- tied to specific (and very old) versions
of those apps. Better for the user to make sure they're in the PATH.