Add Text::Markup::CommonMark
This module, when loaded, will replace Text::Markup::Markdown for
parsing Markdown documents. The default behavior is to keep using
Text::Markup::Markdown, but this gives people the option to switch just
by installing CommonMark and using this new module.

Testing CommonMark on Linux and macOS, but not yet on Windows, as I
haven't figured out how to build CommonMark on Windows, although I did
try in nwellnhof/perl-commonmark#12.

In order to ensure that Pod tests only run when wanted, move them to xt.
Also sort the list of spelling stop words.