Add Asciidoctor module and refactor command utils
Remove the preference for `asciidoctor` from Text::Markup::Asciidoc and
put it in a new parser, Text::Markup::Asciidoctor. Like
Text::Markup::CommonMark, this module is not auto-loaded by Text::Markup
when it recognizes an AsciiDoc document; it still defaults to
Text::Markup::Asciidoc for backward compatibility. But when the user
uses the module, it will be used in preference to
Text::Markup::Asciidoc.

This makes three parsers that use external command line utilities, so
extract the functions used to find and execute those utilities into a
new module, Text::Markup::Cmd. This reduces a ton of duplicate code and
simplifies adding new CLI-oriented parsers in the future.

It also simplifies the testing, as there is no longer magic behavior in
the Asciidoc and Markdown modules, so each can be tested individually.