Fix and document parse method parameters
They're passed via the `options` array reference parameter to `parse`,
but were usually treated as a hash reference. This worked, since most of
the parsers were looking for a `raw` parameter and just passing anything
else, and likely no one has ever passed anything.

But make the behavior match the documentation, and expand the
documentation for each parser with a correct example and a list of some
or all supported options.