Parse Markdown with Text::Markdown::Discount
Replace use of Text::Markdown with Text::Markdown::Discount, as it is
better maintained and supports tables, definition lists, code fences,
and more. Add tests for these features, as well.

This does mean that the old Text::Markdown options will be ignored, so
remove them from the documentation, as well.

While at it:

*   Fix old Changes item and Pod that incorrectly referred to
    Text::Markdown instead of Text::Markup
*   Fix old Pod links and incorrectly linking to Text::Markup instead of
    Parse::BBCode and Text::WikiCreole
*   Update copyright year
*   Test on Perl 5.42, but skip broken Windows Perl builds
*   Increment to v0.40 and timestamp for release